🚚 폴더구조 정리 & 불필요한 코드 삭제
