+++ src/component/custom/customer/shop/customer-shop/shop-content/shop-information/shop-aom/aom-skelton/index.tsx
