+++ src/component/custom/customer/shop/customer-shop/shop-content/shop-information/shop-aom/index.tsx
