+++ src/component/custom/customer/shop/customer-shop/shop-content/shop-information/shop-details/info-card/index.tsx
