+++ src/component/custom/customer/shop/customer-shop/shop-content/shop-information/shop-details/notice/index.tsx
