+++ src/component/custom/customer/shop/customer-shop/shop-content/shop-information/shop-details/price/index.tsx
