+++ src/component/custom/customer/shop/customer-shop/shop-content/shop-information/shop-details/work-hour/work-hour-modal/daily-opening-hours/daily-opening-hours.util.ts
